Dirty Reviews
Gives us the amount of times the word "dirty" was left in the review. It will aslo give us the AirBnB hosts name and link to their reviews on AirBnB. This will be a great way to promote a cleaning service to help with their overall reviews and improve satisfaction moving forward.